retina_live_agent/requirements.txt
# ── Core AI / Agent frameworks ─────────────────────────────────────────────────
google-adk>=1.0.0
google-genai>=1.0.0

# ── MCP (Model Context Protocol) ───────────────────────────────────────────────
mcp[cli]>=1.3.0

# ── Vision & streaming ─────────────────────────────────────────────────────────
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0

# ── ML / model inference ───────────────────────────────────────────────────────
tensorflow>=2.15.0          # or torch if you retrain on PyTorch
# torch>=2.2.0              # uncomment if using PyTorch
# torchvision>=0.17.0

# ── Backend API ────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9     # file uploads

# ── Dashboard ──────────────────────────────────────────────────────────────────
streamlit>=1.33.0

# ── Utilities ──────────────────────────────────────────────────────────────────
audio-recorder-streamlit>=0.0.8
python-dotenv>=1.0.0
pydantic>=2.6.0
